docker/airflow/requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.7
scikit-learn==1.7.2
mlflow==3.6.0
feast==0.40.0
airflow-provider-mlflow==1.1.0
pydantic>=2.0.0
pyyaml
